+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ Changes:
 
 Fixes:
 * handle missing configuration parameters for Ultima
+* don't show consumables where the lot number is missing
 
 # 2.7.0